refactor: split pause-point sources below the file-length limit
Extract cohesive helper classes so every PausePoint Unity-side file is at
or under the 500-line limit. Nested RetargetLineDriftWarning is internal
so the Patcher ledger field can still name it. Skip/warning/assert strings
stay byte-identical.
