+++ Packages/src/Editor/FirstPartyTools/PausePoint/PausePointEnableWarnings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10ddeeb0776284f2c97f442ab3bd76e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
