+++ Packages/src/Editor/FirstPartyTools/PausePoint/PausePointUseCase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1bac73c8a39b84cd0b3b746f24a2c27f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
