+++ Packages/src/Editor/FirstPartyTools/PausePoint/SourcePausePointCaptureEligibility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13481f8aa836944f59bcc6e4da9b8729
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
