+++ Packages/src/Editor/FirstPartyTools/PausePoint/SourcePausePointHotReloadRetarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 899e58273a62748ce8dbd031e96d23f7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
