+++ Packages/src/Editor/FirstPartyTools/PausePoint/SourcePausePointInjectionEmitter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 981746ee425e9472ea92840e9f55f572
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
